Add Gondo's Upgrades as Progressive Items (#213)
* Add progressive items

* Switch Bow, Bug Net, and Slingshot to progressive

* Add gondo upgrade storyflags

* My line breaks are broken

* Update changedflags.md

* Add progressive gondo upgrades

* Add macros for gondo upgrades

* Add debug prerandomization

* Fix wrong text references

* Removed gondo rerandomization 

Github is not a good place to do debug changes.  I will refrain from that

* Add option for scrap shop item shuffle

I tried to make it clear by the title that the items are placed rather than the upgrades at the scrap shop giving something random.

* Add functionality for upgrade toggle

It's near the top but only because I was using how the Rupoors are added to the pool to figure out how to remove progress items.

* Add a couple logical uses for the Swift Beetle

Added Swift Beetle as a potential access to Secret Passage and Crystal on Waterfall Island

* Updated a name to fit the UI better

* Add scrap shop upgrades to UI

* Syntax fix

Really?  Forgetting an indent?

* Fix programming error

* Update logic.py

* Add files via upload

* Added a Tough Beetle macro for Secret Passageway

Didn't know the Heetle trick for Secret Passageway wasn't actually any easier with the swift beetle.  This and any other future long distance bomb carries afforded by the upgrades would need Tough Beetle as a requirement, not Swift.

* Update checks.yaml

* Update SS Rando Logic - Glitchless Requirements.yaml

Changed Swift Beetle to Quick Beetle

* CRLF TO LF

* merge fixes

* Fix broken UI merge and upgrades-off queetle logic

* Made on Accident

Made on accident while trying to generate UI files in the wrong directory

* Run Black

* Remove pointless settings check

* Update SS Rando Logic - Glitched Requirements.yaml

Co-authored-by: lepelog <25211966+lepelog@users.noreply.github.com>

Co-authored-by: lepelog <25211966+lepelog@users.noreply.github.com>